Fixed #102: Add receiver parameter support
This was mostly in place, but since JDKs are generally
compiled without parameter metadata support, most methods
were matching a supplied parameter query of `this`.  (Juggle
was ignoring parameter names where candidates didn't have
metadata.)

Solution: special-case a query for `this`; Juggle now requires
such a parameter name to match, which it should for receiver
parameters since the name is synthesised when creating the
Candidate object.